Marcell

+++ src/main/java/com/example/m08/Laporan/LaporanRepository.java
@@ -5,10 +5,11 @@ import org.springframework.data.jpa.repository.Query;
 import java.util.List;
 import java.time.LocalDate;
 
-public interface LaporanRepository extends JpaRepository<Laporan, Long> {
+public interface LaporanRepository extends JpaRepository<Laporan, Integer> {
     List<Laporan> findByRentDateBetween(LocalDate startDate, LocalDate endDate);
     
     @Query("SELECT l FROM Laporan l WHERE EXTRACT(MONTH FROM l.rentDate) = ?1 AND EXTRACT(YEAR FROM l.rentDate) = ?2")
-List<Laporan> findByMonthAndYear(int month, int year);
+    List<Laporan> findByMonthAndYear(int month, int year);
+    
     Laporan findByIdSewa(int idSewa);
 }
